+++ internal/cmd/root.go
@@ -45,7 +45,7 @@ func InitCmd() {
 
 	managers.InitManagerConfig()
 	config.Refresh()
-	mf := managers.InitManagerFactory()
+	mf := managers.InitManagerFactory(managers.ManagersRegistered, true)
 
 	for _, mName := range mf.ListManagers() {
 		manager, err := mf.GetManager(mName)
@@ -99,12 +99,13 @@ func InitCmd() {
 
 	config.CheckConfig()
 
+	config.AssumeYes = rootCmd.PersistentFlags().BoolP("assumeyes", "y", false, "Automatically answer yes for all questions")
+
 	// cobra.OnInitialize(initConfig)
+	// rootCmd.PersistentFlags().StringP("author", "a", "YOUR NAME", "Author name for copyright attribution")
 	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.cobra.yaml)")
 	// rootCmd.PersistentFlags().StringVarP(&projectBase, "projectbase", "b", "", "base project directory eg. github.com/spf13/")
-	// rootCmd.PersistentFlags().StringP("author", "a", "YOUR NAME", "Author name for copyright attribution")
 	// rootCmd.PersistentFlags().StringVarP(&userLicense, "license", "l", "", "Name of license for the project (can provide `licensetext` in config)")
-	// rootCmd.PersistentFlags().Bool("viper", true, "Use Viper for configuration")
 	// viper.BindPFlag("author", rootCmd.PersistentFlags().Lookup("author"))
 	// viper.BindPFlag("projectbase", rootCmd.PersistentFlags().Lookup("projectbase"))
 	// viper.BindPFlag("useViper", rootCmd.PersistentFlags().Lookup("viper"))